Replace LONG and ULONG with s32 and u32, respectively
These were especially bad because they weren't even defined
as 'long' and 'unsigned long' -- they were defined as 32
bits even on 64-bit platforms.

+++ Parser.cpp
@@ -39,7 +39,7 @@ Parser::~Parser()
 *  Returns:
 ******************************************************************************/
 void Parser::parseText(pp_context *pcontext, ExecBuilder *builder, LPSTR scriptText,
-                      ULONG startingLineNumber, LPCSTR path)
+                      int startingLineNumber, LPCSTR path)
 {
     this->execBuilder = builder;
