+++ ScriptVariant.c
@@ -59,7 +59,7 @@ void ScriptVariant_ParseStringConstant(ScriptVariant *var, char *str)
     StrCache_Copy(var->strVal, str);
 }
 
-HRESULT ScriptVariant_IntegerValue(ScriptVariant *var, LONG *pVal)
+HRESULT ScriptVariant_IntegerValue(ScriptVariant *var, s32 *pVal)
 {
     if(var->vt == VT_INTEGER)
     {
@@ -67,7 +67,7 @@ HRESULT ScriptVariant_IntegerValue(ScriptVariant *var, LONG *pVal)
     }
     else if(var->vt == VT_DECIMAL)
     {
-        *pVal = (LONG)var->dblVal;
+        *pVal = (s32)var->dblVal;
     }
     else
     {
@@ -232,7 +232,7 @@ ScriptVariant *ScriptVariant_And( ScriptVariant *svar, ScriptVariant *rightChild
 ScriptVariant *ScriptVariant_Bit_Or( ScriptVariant *svar, ScriptVariant *rightChild )
 {
     static ScriptVariant retvar = {{.ptrVal = NULL}, VT_EMPTY};
-    LONG l1, l2;
+    s32 l1, l2;
     if(ScriptVariant_IntegerValue(svar, &l1) == S_OK &&
             ScriptVariant_IntegerValue(rightChild, &l2) == S_OK)
     {
@@ -250,7 +250,7 @@ ScriptVariant *ScriptVariant_Bit_Or( ScriptVariant *svar, ScriptVariant *rightCh
 ScriptVariant *ScriptVariant_Xor( ScriptVariant *svar, ScriptVariant *rightChild )
 {
     static ScriptVariant retvar = {{.ptrVal = NULL}, VT_EMPTY};
-    LONG l1, l2;
+    s32 l1, l2;
     if(ScriptVariant_IntegerValue(svar, &l1) == S_OK &&
             ScriptVariant_IntegerValue(rightChild, &l2) == S_OK)
     {
@@ -268,7 +268,7 @@ ScriptVariant *ScriptVariant_Xor( ScriptVariant *svar, ScriptVariant *rightChild
 ScriptVariant *ScriptVariant_Bit_And( ScriptVariant *svar, ScriptVariant *rightChild )
 {
     static ScriptVariant retvar = {{.ptrVal = NULL}, VT_EMPTY};
-    LONG l1, l2;
+    s32 l1, l2;
     if(ScriptVariant_IntegerValue(svar, &l1) == S_OK &&
             ScriptVariant_IntegerValue(rightChild, &l2) == S_OK)
     {
@@ -474,12 +474,12 @@ ScriptVariant *ScriptVariant_Le( ScriptVariant *svar, ScriptVariant *rightChild 
 ScriptVariant *ScriptVariant_Shl( ScriptVariant *svar, ScriptVariant *rightChild )
 {
     static ScriptVariant retvar = {{.ptrVal = NULL}, VT_EMPTY};
-    LONG l1, l2;
+    s32 l1, l2;
     if(ScriptVariant_IntegerValue(svar, &l1) == S_OK &&
             ScriptVariant_IntegerValue(rightChild, &l2) == S_OK)
     {
         retvar.vt = VT_INTEGER;
-        retvar.lVal = ((ULONG)l1) << ((ULONG)l2);
+        retvar.lVal = ((u32)l1) << ((u32)l2);
     }
     else
     {
@@ -493,12 +493,12 @@ ScriptVariant *ScriptVariant_Shl( ScriptVariant *svar, ScriptVariant *rightChild
 ScriptVariant *ScriptVariant_Shr( ScriptVariant *svar, ScriptVariant *rightChild )
 {
     static ScriptVariant retvar = {{.ptrVal = NULL}, VT_EMPTY};
-    LONG l1, l2;
+    s32 l1, l2;
     if(ScriptVariant_IntegerValue(svar, &l1) == S_OK &&
             ScriptVariant_IntegerValue(rightChild, &l2) == S_OK)
     {
         retvar.vt = VT_INTEGER;
-        retvar.lVal = ((ULONG)l1) >> ((ULONG)l2);
+        retvar.lVal = ((u32)l1) >> ((u32)l2);
     }
     else
     {
@@ -525,7 +525,7 @@ ScriptVariant *ScriptVariant_Add( ScriptVariant *svar, ScriptVariant *rightChild
         else
         {
             ScriptVariant_ChangeType(&retvar, VT_INTEGER);
-            retvar.lVal = (LONG)(dbl1 + dbl2);
+            retvar.lVal = (s32)(dbl1 + dbl2);
         }
     }
     else if(svar->vt == VT_STR || rightChild->vt == VT_STR)
@@ -559,7 +559,7 @@ ScriptVariant *ScriptVariant_Sub( ScriptVariant *svar, ScriptVariant *rightChild
         else
         {
             retvar.vt = VT_INTEGER;
-            retvar.lVal = (LONG)(dbl1 - dbl2);
+            retvar.lVal = (s32)(dbl1 - dbl2);
         }
     }
     else
@@ -586,7 +586,7 @@ ScriptVariant *ScriptVariant_Mul( ScriptVariant *svar, ScriptVariant *rightChild
         else
         {
             retvar.vt = VT_INTEGER;
-            retvar.lVal = (LONG)(dbl1 * dbl2);
+            retvar.lVal = (s32)(dbl1 * dbl2);
         }
     }
     else
@@ -617,7 +617,7 @@ ScriptVariant *ScriptVariant_Div( ScriptVariant *svar, ScriptVariant *rightChild
         else
         {
             retvar.vt = VT_INTEGER;
-            retvar.lVal = (LONG)(dbl1 / dbl2);
+            retvar.lVal = (s32)(dbl1 / dbl2);
         }
     }
     else
@@ -632,7 +632,7 @@ ScriptVariant *ScriptVariant_Div( ScriptVariant *svar, ScriptVariant *rightChild
 ScriptVariant *ScriptVariant_Mod( ScriptVariant *svar, ScriptVariant *rightChild )
 {
     static ScriptVariant retvar = {{.ptrVal = NULL}, VT_EMPTY};
-    LONG l1, l2;
+    s32 l1, l2;
     if(ScriptVariant_IntegerValue(svar, &l1) == S_OK &&
             ScriptVariant_IntegerValue(rightChild, &l2) == S_OK)
     {
@@ -789,7 +789,7 @@ ScriptVariant *ScriptVariant_Boolean_Not(ScriptVariant *svar )
 ScriptVariant *ScriptVariant_Bit_Not(ScriptVariant *svar )
 {
     static ScriptVariant retvar = {{.lVal=0}, VT_INTEGER};
-    LONG l1;
+    s32 l1;
     if(ScriptVariant_IntegerValue(svar, &l1) == S_OK)
     {
         retvar.lVal = ~l1;
